Moved ToHeapDataConverter to utility package
